+++ commands/image_mosh.js
@@ -1,0 +1,34 @@
+const { readPNG, moshImage } = require('../utils/util')
+const Eris = require('eris')
+const fs = require('fs')
+
+const image_mosh = new Eris.Command('dm', async (msg, args) => {
+    const bot = require('../discord-bot')
+    if (msg.attachments[0]) {
+        const attachment = msg.attachments[0]
+        let dataMode
+        if (args.length > 0) {
+            dataMode = args[0]
+        }
+        try {
+            await bot.deleteMessage(msg.channel.id, msg.id)
+            await readPNG(attachment.url, attachment.id)
+            await moshImage(attachment.id, dataMode)
+            await bot.createMessage(msg.channel.id, 'Moshed Image!', {
+                file: fs.readFileSync(`images/moshed_${attachment.id}.jpg`),
+                name: `moshed_${attachment.id}.jpg`
+            })
+            console.log('File sent!')
+            fs.unlinkSync(`images/moshed_${attachment.id}.jpg`)
+        } catch (e) {
+            console.log(e)
+            bot.createMessage(msg.channel.id, e)
+        }
+    }
+    else return 'Invalid input.'
+}, {
+    description: 'Use Datamosh to mosh a lame image into a SCHIFTY pic!',
+
+})
+
+module.exports = image_mosh
